fixing auth colors

+++ source_code/app/src/main/java/com/fola/habit_tracker/data/database/Entities.kt
@@ -50,7 +50,7 @@ data class DailyHabits(
     @ColumnInfo(name = "habit_id") val id: Long = 0,
     @ColumnInfo(name = "is_done") val isDone: Boolean = false,
 
-    val progress: Float = 0f,
+    val progress: Int = 0,
 )
 
 data class DayWithHabits(
